// src/lib/types/AssetEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AssetEvent = { id: number, accountId: number | null, assetLabel: string | null, date: string, description: string, kind: string, cost: number, assetValue: number | null, vendor: string | null, notes: string | null, lifeExpectancy: string | null, linkedTransactionId: number | null, createdAt: string, updatedAt: string, hasAttachment: boolean, };